feat(constants): impl fork policy for block versions
