+++ constants/src/lib.rs
@@ -1,5 +1,9 @@
 //* Chain constants.
 
+use block_version::BlockVersion;
+
+pub mod block_version;
+
 pub const CURRENT_BLOCK_VERSION: BlockVersion = BlockVersion::GreatVoyage4_0_1;
 
 /// Will postpone txns if block size exceeds 2MiB.
@@ -82,58 +86,3 @@ pub const ADAPTIVE_ENERGY_DECREASE_RATE_DENOMINATOR: i64 = 100;
 
 pub const ADAPTIVE_ENERGY_INCREASE_RATE_NUMERATOR: i64 = 1000;
 pub const ADAPTIVE_ENERGY_INCREASE_RATE_DENOMINATOR: i64 = 999;
-
-/// Block versions. These versions match version names on github release page(or PR numbers).
-#[derive(PartialEq, Eq, PartialOrd, Ord, Debug)]
-pub enum BlockVersion {
-    /// Also applies to all blocks before around #2300000.
-    Genesis = 0,
-    Unknown1290 = 1,
-    // PR #1442
-    Odyssey3_0_1 = 2,
-    // PR #1485
-    Odyssey3_1_0 = 3,
-    /// Special Check after block #4727890 (enery.limit.block.num).
-    /// When block.version == 5,
-    /// it makes block use new energy to handle transaction when block number >= 4727890L.
-    /// Otherwise version !=5, skip.
-    ///
-    /// - Support Of Resource Delegation.
-    /// - UpdateEnergyLimitContract
-    /// - TotalEnergyLimit
-    ///
-    /// Renamed: ENERGY_LIMIT
-    Odyssey3_2 = 5,
-    /// - Deprecates `TotalEnergyLimit`
-    /// - Add `TotalCurrentEnergyLimit`
-    Odyssey3_2_2 = 6,
-    /// - AllowMultisig
-    /// - AllowAdaptiveEnergy
-    /// - UpdateAccountPermissionFee
-    /// - MultisigFee
-    Odyssey3_5 = 7,
-    /// - AllowProtoFilterNum
-    /// - AllowAccountStateRoot
-    /// - AllowTvmConstantinopleUpgrade
-    Odyssey3_6_0 = 8,
-    /// - AllowTvmSolidity059Upgrade
-    /// - AdaptiveResourceLimitTargetRatio
-    /// - AdaptiveResourceLimitMultiplier
-    /// - AllowChangeDelegation
-    /// - StandbyWitnessPayPerBlock
-    Odyssey3_6_5 = 9,
-    /// - ForbidTransferToContract
-    Odyssey3_6_6 = 10,
-    // Note: This version has only non-core API changes.
-    Odyssey3_7 = 15,
-    /// Shielded TVM precompiles.
-    ///
-    /// - support AllowTvmShieldedUpgrade config
-    GreatVoyage4_0_0 = 16,
-    /// First hard fork based on timestamp. Fork at 1596780000_000, at least 22 SRs.
-    ///
-    /// See-also: https://github.com/tronprotocol/java-tron/pull/3314
-    ///
-    /// - support AllowTvmShieldedUpgrade proposal
-    GreatVoyage4_0_1 = 17,
-}
